+++ tests/tests/chain.rs
@@ -1,6 +1,7 @@
 use std::{path::PathBuf, sync::Arc};
 
 use alloy::hex::ToHexExt;
+use alloy::network::AnyNetwork;
 use alloy::providers::{DynProvider, Provider};
 use alloy::sol;
 use alloy::{
@@ -29,7 +30,7 @@ sol!(
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
-type StateT = Option<(Arc<AnvilInstance>, DynProvider)>;
+type StateT = Option<(Arc<AnvilInstance>, DynProvider<AnyNetwork>)>;
 
 static TEST_CHAIN_STATE: Mutex<StateT> = Mutex::const_new(None);
 
@@ -38,7 +39,7 @@ static TEST_CHAIN_STATE: Mutex<StateT> = Mutex::const_new(None);
 #[allow(dead_code)]
 pub struct TestChain<'a> {
     pub anvil: Arc<AnvilInstance>,
-    pub rpc_client: DynProvider,
+    pub rpc_client: DynProvider<AnyNetwork>,
     // Anvil does not like concurrent access so we serialize
     // all tests that are accessing it
     guard: MutexGuard<'a, StateT>,
@@ -51,7 +52,7 @@ pub async fn get_test_chain() -> TestChain<'static> {
 
     if state_guard.is_none() {
         let anvil = Anvil::new().spawn();
-        let rpc_client = ProviderBuilder::new()
+        let rpc_client = ProviderBuilder::new_with_network::<AnyNetwork>()
             .connect(&anvil.endpoint())
             .await
             .unwrap()